feat: fix highlighted text foreground color.
Issue
---
When using the theme the foreground color of the selected text is too
similar to the background color set by logseq on selection. This is bad
from an accessabilty standpoint and makes using logseq with this theme
harder due to missing contrast.

Background
---
logseq-catppuccin sets the foreground color only while leaving it to the
logseq default theme to handle the background color.

Solution
---
I looked at how the obsidian catppuccin theme handled the color of the
selection. They introduced an new var `--text-on-accent` that handles
the foreground color. It is basically just a new name for the value in
`--ctp-base`. I added a new var `--ls-text-on-accent` recreating this
behavior and use this as the value for the selection's foreground color.